contract Balance adjustments updates to total/max supply and balances added

+++ contracts/contracts/layer_2/interfaces/equity/IEquity.sol
@@ -265,7 +265,7 @@ interface IEquity {
     struct BalanceAdjustment {
         uint256 executionDate;
         uint256 factor;
-        uint256 decimals;
+        uint8 decimals;
     }
 
     function getEquityDetails()
